services/ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
boto3==1.34.131
sentence-transformers==2.2.2
xgboost==2.0.3
numpy==1.26.4
scikit-learn==1.4.2
pdfminer.six==20231228
